Refactor FennelSymbol pattern
Remove `&` symbol from it, as it is not allowed in recent versions of
Fennel.
Make highlighting of FennelSymbol optional (on by default). Motivation
to make this because it just highlight almost everything previously not
highlighted, in my opinion it is work for Normal highlight group. And
this pattern is the most cpu time consuming from all syntax file.